requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp_socks==0.9.2
aiosignal==1.3.2
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
frozenlist==1.5.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matrix-nio==0.25.2
multidict==6.1.0
propcache==0.2.1
pycryptodome==3.21.0
python-socks==2.5.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
unpaddedbase64==2.1.0
urllib3==2.2.3
yarl==1.18.3
